Refactor: Improve MangaDex API integration and UI components
This commit introduces several improvements related to MangaDex API usage and associated UI components:

- **Remove `MangaProvider` base class:** The abstract `MangaProvider` class is removed as it was not being used effectively and the client instance could be handled directly within `MangaDexAPI`.
- **Centralize MangaDex types:** Created `MangaDexTypes.kt` to hold common data classes like `Manga`, `ChapterModel`, `MangaTag`, `MangaDetailedState`, and `SearchState`.
- **Update `ChaptersRepository` and `ReadMangaScreen`:** Updated references to `ChapterModel` to point to the new centralized `MangaDexTypes`.
- **Enhance `MangaDexAPI`:**
    - Increased the search result limit from 20 to 50.
    - Added `getMangaDetails` function to fetch detailed information for a specific manga ID, including description, status, year, cover art, and tags.
    - Added `getChapters` function to fetch all chapters for a manga ID, including translation group information and language flags. Chapters are grouped and sorted by number, with English translations prioritized.
    - Added a helper function `getLanguageFlag` to provide emoji flags for various languages.
- **Update `MangaDetailedViewModel`:**
    - Refactored the ViewModel to use the new `getMangaDetails` and `getChapters` functions from `MangaDexAPI` instead of the previous fragmented approach.
    - Updated the internal state (`_uiState`) to use the new `MangaDetailedState` data class.
- **Update `SearchViewModel`:**
    - Adjusted the parsing of MangaDex search results to align with the updated API response structure, particularly for tags.
- **Improve `AnimeDetailedScreen` layout:**
    - Fixed the wide screen layout issue where content wasn't properly laid out side-by-side by using a `Column` as the main container and a `Row` for side-by-side elements (charts and trailer).
    - Standardized the trailer height for better consistency.
- **Update `MangaDetailedScreen` and `SearchScreen`:** Updated references to `MangaTag` to point to the new centralized `MangaDexTypes` and adjusted how tag names are accessed.
