--- app/src/main/java/me/thuanc177/kotsune/libs/MangaProvider.kt
@@ -1,21 +1,0 @@
-package me.thuanc177.kotsune.libs
-
-import okhttp3.OkHttpClient
-import java.util.concurrent.TimeUnit
-
-abstract class MangaProvider {
-    protected val client: OkHttpClient = OkHttpClient.Builder()
-        .connectTimeout(15, TimeUnit.SECONDS)
-        .readTimeout(15, TimeUnit.SECONDS)
-        .addInterceptor { chain ->
-            val request = chain.request().newBuilder()
-                .header("User-Agent", USER_AGENT)
-                .build()
-            chain.proceed(request)
-        }
-        .build()
-
-    companion object {
-        private const val USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.110 Safari/537.36"
-    }
-}
