+++ app/src/main/java/me/thuanc177/kotsune/libs/mangaProvider/mangadex/MangaDexAPI.kt
@@ -4,19 +4,21 @@ import android.util.Log
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import me.thuanc177.kotsune.config.AppConfig
-import me.thuanc177.kotsune.libs.MangaProvider
 import me.thuanc177.kotsune.libs.common.fetchMangaInfoFromBal
-import okhttp3.Request
+import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.ChapterModel
+import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.Manga
+import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.MangaTag
 import org.json.JSONArray
 import java.net.URL
 import org.json.JSONObject
+import java.io.IOException
 import java.net.HttpURLConnection
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
 class MangaDexAPI (
     private val appConfig: AppConfig
-): MangaProvider() {
+) {
     private val TAG = "MangaDexAPI"
     private val baseUrl = "https://api.mangadex.org"
 
@@ -87,7 +89,7 @@ class MangaDexAPI (
         try {
             val encodedTitle = java.net.URLEncoder.encode(title, "UTF-8")
             val contentRatingParams = getContentRatingParams()
-            val url = URL("$baseUrl/manga?title=$encodedTitle&limit=20&includes[]=cover_art&includes[]=author&includes[]=rating$contentRatingParams")
+            val url = URL("$baseUrl/manga?title=$encodedTitle&limit=50&includes[]=cover_art&includes[]=author&includes[]=rating$contentRatingParams")
 
             Log.d(TAG, "Searching manga with URL: $url")
             val connection = url.openConnection() as HttpURLConnection
@@ -156,26 +158,29 @@ class MangaDexAPI (
                         Log.w(TAG, "No cover art found for manga ID: $mangaId")
                     }
 
-                    // Get tags
-                    val tags = mutableListOf<String>()
+                    val tags = mutableListOf<MangaTag>()
                     val tagsArray = attributes.getJSONArray("tags")
                     for (j in 0 until tagsArray.length()) {
                         val tag = tagsArray.getJSONObject(j)
+                        val tagId = tag.getString("id")
                         val tagAttributes = tag.getJSONObject("attributes")
                         val tagName = tagAttributes.getJSONObject("name")
-                        tagName.optString("en")?.let { tags.add(it) }
+                        val nameEn = tagName.optString("en")
+                        if (!nameEn.isNullOrEmpty()) {
+                            tags.add(MangaTag(id = tagId, tagName = nameEn))
+                        }
                     }
+
 
                     val mangaMap = mapOf(
                         "id" to mangaId,
                         "title" to title,
-                        "poster" to (coverImage ?: ""), // Use empty string instead of null
+                        "poster" to (coverImage ?: ""),
                         "type" to status,
                         "genres" to tags,
                         "year" to year,
                         "contentRating" to contentRating
                     )
-
                     results.add(mangaMap as Map<String, Any>)
                 }
 
@@ -214,64 +219,305 @@ class MangaDexAPI (
         return contentRatingBuilder.toString()
     }
 
-    // Existing methods remain unchanged
-    suspend fun getManga(mangadexMangaId: String): Map<String, Any?>? = withContext(Dispatchers.IO) {
-        val balData = fetchMangaInfoFromBal(mangadexMangaId) ?: return@withContext null
-
-        val sitesObj = balData.getJSONObject("Sites")
-        val mangadexObj = sitesObj.getJSONObject("Mangadex")
-        val mangaId = mangadexObj.keys().next()
-        val mangaDexManga = mangadexObj.getJSONObject(mangaId)
-
-        return@withContext mapOf(
-            "id" to mangaId,
-            "title" to mangaDexManga.getString("title"),
-            "poster" to mangaDexManga.getString("image"),
-            "availableChapters" to emptyList<String>()
-        )
-    }
-
-    suspend fun getChapterThumbnails(mangadexMangaId: String, chapter: String): Map<String, Any>? = withContext(Dispatchers.IO) {
+    suspend fun getMangaDetails(mangaId: String): Manga? = withContext(Dispatchers.IO) {
         try {
-            val chapterInfoUrl = "https://api.mangadex.org/chapter?manga=$mangadexMangaId&translatedLanguage[]=en&chapter=$chapter&includeEmptyPages=0"
-            val chapterInfoRequest = Request.Builder().url(chapterInfoUrl).build()
+            Log.d(TAG, "Fetching manga details for ID: $mangaId")
+            val url = "https://api.mangadex.org/manga/$mangaId?includes[]=cover_art&includes[]=author&includes[]=artist&includes[]=tag"
+            val connection = URL(url).openConnection() as HttpURLConnection
+            connection.requestMethod = "GET"
+            connection.connectTimeout = 15000
+            connection.readTimeout = 15000
 
-            val chapterInfoResponse = client.newCall(chapterInfoRequest).execute()
-            if (!chapterInfoResponse.isSuccessful) return@withContext null
+            if (connection.responseCode == 200) {
+                val response = connection.inputStream.bufferedReader().use { it.readText() }
+                val jsonObject = JSONObject(response)
+                val data = jsonObject.getJSONObject("data")
+                val attributes = data.getJSONObject("attributes")
 
-            val chapterInfoJson = JSONObject(chapterInfoResponse.body?.string() ?: "")
-            val dataArray = chapterInfoJson.getJSONArray("data")
-            if (dataArray.length() == 0) return@withContext null
+                // Extract title
+                val titleObj = attributes.getJSONObject("title")
+                val titles = mutableListOf<String>()
+                for (key in titleObj.keys()) {
+                    titleObj.getString(key)?.let { titles.add(it) }
+                }
 
-            val chapterInfo = dataArray.getJSONObject(0)
-            val chapterId = chapterInfo.getString("id")
+                // Extract description
+                val description = attributes.optJSONObject("description")?.let { descObj ->
+                    descObj.optString("en") ?: run {
+                        // If English description not available, get the first available description
+                        for (key in descObj.keys()) {
+                            val desc = descObj.optString(key)
+                            if (desc.isNotEmpty()) return@run desc
+                        }
+                        ""
+                    }
+                } ?: ""
 
-            val thumbnailsUrl = "https://api.mangadex.org/at-home/server/$chapterId"
-            val thumbnailsRequest = Request.Builder().url(thumbnailsUrl).build()
-            val thumbnailsResponse = client.newCall(thumbnailsRequest).execute()
+                val status = attributes.optString("status", "unknown")
+                val year = attributes.optInt("year", 0)
+                val contentRating = attributes.optString("contentRating", "safe")
 
-            if (!thumbnailsResponse.isSuccessful) return@withContext null
+                // Get cover image
+                var coverImage: String? = null
+                val relationships = data.getJSONArray("relationships")
+                for (i in 0 until relationships.length()) {
+                    val rel = relationships.getJSONObject(i)
+                    if (rel.getString("type") == "cover_art") {
+                        if (rel.has("attributes")) {
+                            val coverAttributes = rel.getJSONObject("attributes")
+                            val fileName = coverAttributes.optString("fileName")
+                            if (fileName.isNotEmpty()) {
+                                coverImage = "https://uploads.mangadex.org/covers/$mangaId/$fileName"
+                            }
+                        }
+                    }
+                }
 
-            val thumbnailsJson = JSONObject(thumbnailsResponse.body?.string() ?: "")
-            val baseUrl = thumbnailsJson.getString("baseUrl")
-            val chapterObj = thumbnailsJson.getJSONObject("chapter")
-            val hash = chapterObj.getString("hash")
+                // Extract tags
+                val tagsArray = attributes.getJSONArray("tags")
+                val tags = mutableListOf<MangaTag>()
+                for (i in 0 until tagsArray.length()) {
+                    val tag = tagsArray.getJSONObject(i)
+                    val tagId = tag.getString("id")
 
-            val thumbnails = ArrayList<String>()
-            for (i in 0 until dataArray.length()) {
-                thumbnails.add("$baseUrl/data/$hash/${dataArray.getString(i)}")
+                    // Extract tag name from attributes
+                    val tagAttributes = tag.getJSONObject("attributes")
+                    val tagNameObj = tagAttributes.getJSONObject("name")
+
+                    // Try to get English name first, then fallback to any available language
+                    val tagName = tagNameObj.optString("en") ?: run {
+                        var name = ""
+                        val keys = tagNameObj.keys()
+                        while (keys.hasNext()) {
+                            val key = keys.next()
+                            val localizedName = tagNameObj.optString(key)
+                            if (localizedName.isNotEmpty()) {
+                                name = localizedName
+                                break
+                            }
+                        }
+                        name
+                    }
+
+                    // Only add tags with non-empty names
+                    if (tagName.isNotEmpty()) {
+                        tags.add(MangaTag(tagId, tagName))
+                    }
+                }
+
+                return@withContext Manga(
+                    id = mangaId,
+                    title = titles,
+                    poster = coverImage,
+                    status = status,
+                    description = description,
+                    lastUpdated = attributes.optString("updatedAt"),
+                    lastChapter = null,
+                    latestUploadedChapterId = attributes.optString("latestUploadedChapter"),
+                    year = if (year > 0) year else null,
+                    contentRating = contentRating,
+                    tags = tags
+                )
             }
 
-            val attributes = chapterInfo.getJSONObject("attributes")
-            val title = attributes.getString("title")
+            // Handle error responses
+            val errorMessage = when(connection.responseCode) {
+                404 -> "Manga not found"
+                429 -> "Rate limit exceeded. Please try again later."
+                500, 502, 503, 504 -> "Server error. Please try again later."
+                else -> "Error ${connection.responseCode}: ${connection.responseMessage}"
+            }
 
-            return@withContext mapOf(
-                "thumbnails" to thumbnails,
-                "title" to title
-            )
+            Log.e(TAG, "API error: $errorMessage")
+            throw IOException(errorMessage)
         } catch (e: Exception) {
-            Log.e(TAG, "Error getting chapter thumbnails: ${e.message}")
+            Log.e(TAG, "Error parsing manga details", e)
             return@withContext null
+        }
+    }
+
+    suspend fun getChapters(mangaId: String): List<ChapterModel> = withContext(Dispatchers.IO) {
+        try {
+            Log.d(TAG, "Fetching chapters for manga ID: $mangaId")
+            val url = "https://api.mangadex.org/manga/$mangaId/feed?limit=500&includes[]=scanlation_group"
+            val connection = URL(url).openConnection() as HttpURLConnection
+            connection.requestMethod = "GET"
+            connection.connectTimeout = 15000
+            connection.readTimeout = 15000
+
+            if (connection.responseCode == 200) {
+                val response = connection.inputStream.bufferedReader().use { it.readText() }
+                val jsonObject = JSONObject(response)
+                val dataArray = jsonObject.getJSONArray("data")
+                val chapters = mutableListOf<ChapterModel>()
+
+                // Map to store chapter number to its translations
+                val chapterMap = mutableMapOf<String, MutableList<ChapterModel>>()
+
+                for (i in 0 until dataArray.length()) {
+                    val chapterObject = dataArray.getJSONObject(i)
+                    val chapterId = chapterObject.getString("id")
+                    val attributes = chapterObject.getJSONObject("attributes")
+
+                    val chapterNumber = attributes.optString("chapter", "")
+                    if (chapterNumber.isEmpty()) continue  // Skip chapters without numbers
+
+                    // Get language code
+                    val language = attributes.optString("translatedLanguage", "en")
+
+                    // Get scanlation group
+                    var translatorGroup: String? = null
+                    val relationships = chapterObject.getJSONArray("relationships")
+                    for (j in 0 until relationships.length()) {
+                        val rel = relationships.getJSONObject(j)
+                        if (rel.getString("type") == "scanlation_group" && rel.has("attributes")) {
+                            translatorGroup = rel.getJSONObject("attributes").optString("name")
+                            break
+                        }
+                    }
+
+                    val chapterTitle = attributes.optString("title", "")
+                    val finalTitle = if (chapterTitle.isBlank()) {
+                        "Chapter $chapterNumber"
+                    } else {
+                        chapterTitle
+                    }
+
+                    val chapterModel = ChapterModel(
+                        id = chapterId,
+                        number = chapterNumber,
+                        title = finalTitle,
+                        publishedAt = attributes.optString("publishAt", ""),
+                        pages = attributes.optInt("pages", 0),
+                        volume = attributes.optString("volume", null),
+                        isRead = false, // Will be updated by the ViewModel
+                        language = language,
+                        translatorGroup = translatorGroup,
+                        languageFlag = getLanguageFlag(language)
+                    )
+
+                    // Add to chapter map
+                    if (!chapterMap.containsKey(chapterNumber)) {
+                        chapterMap[chapterNumber] = mutableListOf()
+                    }
+                    chapterMap[chapterNumber]?.add(chapterModel)
+                }
+
+                // Sort each chapter's translations by language (English first)
+                chapterMap.forEach { (_, translations) ->
+                    translations.sortWith(compareBy<ChapterModel> { it.language != "en" }
+                        .thenBy { it.publishedAt }
+                        .thenBy { it.language })
+                }
+
+                // Flatten the map back to a list, sorted by chapter number
+                val sortedChapterNumbers = chapterMap.keys
+                    .sortedWith(compareBy {
+                        it.toFloatOrNull() ?: Float.MAX_VALUE
+                    })
+
+                for (chapterNum in sortedChapterNumbers) {
+                    chapters.addAll(chapterMap[chapterNum] ?: emptyList())
+                }
+
+                return@withContext chapters
+            }
+
+            // Handle error responses
+            val errorMessage = when(connection.responseCode) {
+                404 -> "Chapters not found"
+                429 -> "Rate limit exceeded. Please try again later."
+                500, 502, 503, 504 -> "Server error. Please try again later."
+                else -> "Error ${connection.responseCode}: ${connection.responseMessage}"
+            }
+
+            Log.e(TAG, "API error: $errorMessage")
+            throw IOException(errorMessage)
+        } catch (e: Exception) {
+            Log.e(TAG, "Error fetching chapters", e)
+            return@withContext emptyList()
+        }
+    }
+
+    private fun getLanguageFlag(languageCode: String): String {
+        // Extract the base language and region if present (e.g., "en-us" -> "en" and "us")
+        val parts = languageCode.split("-")
+        if (parts.size > 1) parts[1].uppercase() else ""
+
+        // For languages with specific regions, use the region's flag
+        return when (languageCode) {
+            "en" -> "🇬🇧"  // English (default to UK)
+            "en-us" -> "🇺🇸"  // English (US)
+            "ja" -> "🇯🇵"  // Japanese
+            "ko" -> "🇰🇷"  // Korean
+            "zh" -> "🇨🇳"  // Chinese (simplified)
+            "zh-hk", "zh-tw" -> "🇹🇼"  // Chinese (traditional)
+            "ru" -> "🇷🇺"  // Russian
+            "fr" -> "🇫🇷"  // French
+            "de" -> "🇩🇪"  // German
+            "es" -> "🇪🇸"  // Spanish
+            "es-la" -> "🇲🇽"  // Spanish (Latin America)
+            "pt-br" -> "🇧🇷"  // Portuguese (Brazil)
+            "pt" -> "🇵🇹"  // Portuguese
+            "it" -> "🇮🇹"  // Italian
+            "pl" -> "🇵🇱"  // Polish
+            "tr" -> "🇹🇷"  // Turkish
+            "th" -> "🇹🇭"  // Thai
+            "vi" -> "🇻🇳"  // Vietnamese
+            "id" -> "🇮🇩"  // Indonesian
+            "ar" -> "🇸🇦"  // Arabic
+            "hi" -> "🇮🇳"  // Hindi
+            "bn" -> "🇧🇩"  // Bengali
+            "ms" -> "🇲🇾"  // Malay
+            "fi" -> "🇫🇮"  // Finnish
+            "da" -> "🇩🇰"  // Danish
+            "no" -> "🇳🇴"  // Norwegian
+            "sv" -> "🇸🇪"  // Swedish
+            "cs" -> "🇨🇿"  // Czech
+            "sk" -> "🇸🇰"  // Slovak
+            "hu" -> "🇭🇺"  // Hungarian
+            "ro" -> "🇷🇴"  // Romanian
+            "uk" -> "🇺🇦"  // Ukrainian
+            "bg" -> "🇧🇬"  // Bulgarian
+            "el" -> "🇬🇷"  // Greek
+            "he" -> "🇮🇱"  // Hebrew
+            "lt" -> "🇱🇹"  // Lithuanian
+            "lv" -> "🇱🇻"  // Latvian
+            "et" -> "🇪🇪"  // Estonian
+            "sl" -> "🇸🇮"  // Slovenian
+            "hr" -> "🇭🇷"  // Croatian
+            "sr" -> "🇷🇸"  // Serbian
+            "mk" -> "🇲🇰"  // Macedonian
+            "sq" -> "🇦🇱"  // Albanian
+            "bs" -> "🇧🇦"  // Bosnian
+            "is" -> "🇮🇸"  // Icelandic
+            "ga" -> "🇮🇪"  // Irish
+            "cy" -> "🇬🇧"  // Welsh
+            "eu" -> "🇪🇸"  // Basque
+            "ca" -> "🇪🇸"  // Catalan
+            "sw" -> "🇰🇪"  // Swahili
+            "tl" -> "🇵🇭"  // Tagalog
+            "jw" -> "🇮🇩"  // Javanese
+            "su" -> "🇮🇩"  // Sundanese
+            "la" -> "🇻🇦"  // Latin
+            "tlh" -> "🇰🇷"  // Klingon
+            "xh" -> "🇿🇦"  // Xhosa
+            "zu" -> "🇿🇦"  // Zulu
+            "af" -> "🇿🇦"  // Afrikaans
+            "am" -> "🇪🇹"  // Amharic
+            "hy" -> "🇦🇲"  // Armenian
+            "mn" -> "🇲🇳"  // Mongolian
+            "ne" -> "🇳🇵"  // Nepali
+            "pa" -> "🇮🇳"  // Punjabi
+            "ta" -> "🇮🇳"  // Tamil
+            "te" -> "🇮🇳"  // Telugu
+            "ml" -> "🇮🇳"  // Malayalam
+            "or" -> "🇮🇳"  // Odia
+            "si" -> "🇱🇰"  // Sinhala
+            "my" -> "🇲🇲"  // Burmese
+            "km" -> "🇰🇭"  // Khmer
+            "lo" -> "🇱🇦"  // Lao
+            else -> "🌐 $languageCode"  // Globe and language code
         }
     }
 }
