+++ app/src/main/java/me/thuanc177/kotsune/viewmodel/MangaDetailedViewModel.kt
@@ -4,62 +4,18 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexAPI
-import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.Manga
+import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.ChapterModel
+import me.thuanc177.kotsune.libs.mangaProvider.mangadex.MangaDexTypes.MangaDetailedState
 import me.thuanc177.kotsune.repository.FavoritesRepository
-import org.json.JSONObject
 import java.io.IOException
-import java.net.HttpURLConnection
 import java.net.SocketTimeoutException
-import java.net.URL
 import java.util.concurrent.atomic.AtomicBoolean
-
-data class MangaDetailedState(
-    val manga: Manga? = null,
-    val isLoading: Boolean = false,
-    val error: String? = null,
-    val isFavorite: Boolean = false,
-    val chapters: List<ChapterModel> = emptyList(),
-    val chaptersLoading: Boolean = false,
-    val chaptersError: String? = null,
-    val chapterSortAscending: Boolean = false,
-    val selectedChapterIndex: Int?,
-    val selectedTranslationGroup: String? = null
-)
-
-data class MangaDetailedUiState(
-    val manga: Manga? = null,
-    val chapters: List<ChapterModel> = emptyList(),
-    val isLoading: Boolean = false,
-    val error: String? = null,
-    val chaptersLoading: Boolean = false,
-    val chaptersError: String? = null,
-    val isFavorite: Boolean = false,
-    val chapterSortAscending: Boolean = false, // New sorting direction state
-    val selectedChapterIndex: Int? = null,      // Selected chapter for reading
-    val selectedTranslationGroup: String? = null // Selected translation group preference
-)
-
-data class ChapterModel(
-    val id: String,
-    val number: String,
-    val title: String,
-    val publishedAt: String,
-    val pages: Int = 0,
-    val thumbnail: String? = null,
-    var isRead: Boolean = false,
-    val volume: String? = null,
-    val language: String = "en",
-    val translatorGroup: String? = null,
-    val languageFlag: String? = null
-)
 
 class MangaDetailedViewModel(
     private val mangaDexAPI: MangaDexAPI,
@@ -67,7 +23,7 @@ class MangaDetailedViewModel(
     private val favoritesRepository: FavoritesRepository
 ) : ViewModel() {
 
-   private val _uiState = MutableStateFlow(MangaDetailedState(
+    private val _uiState = MutableStateFlow(MangaDetailedState(
         manga = null,
         isLoading = false,
         error = null,
@@ -102,7 +58,7 @@ class MangaDetailedViewModel(
         viewModelScope.launch {
             _uiState.update { it.copy(isLoading = true, error = null) }
             try {
-                val manga = fetchMangaFromAPI(mangaId)
+                val manga = mangaDexAPI.getMangaDetails(mangaId)
                 if (manga != null) {
                     _uiState.update { it.copy(manga = manga, isLoading = false) }
                     fetchChapters()
@@ -116,6 +72,11 @@ class MangaDetailedViewModel(
                 _uiState.update {
                     it.copy(isLoading = false, error = "Network timeout. Please check your connection.")
                 }
+            } catch (e: Exception) {
+                Log.e(TAG, "Error fetching manga details", e)
+                _uiState.update {
+                    it.copy(isLoading = false, error = e.message ?: "Unknown error occurred")
+                }
             }
         }
     }
@@ -124,12 +85,11 @@ class MangaDetailedViewModel(
         viewModelScope.launch {
             _uiState.update { it.copy(chaptersLoading = true, chaptersError = null) }
             try {
-                val chapters = fetchChaptersFromAPI(mangaId)
+                val chapters = mangaDexAPI.getChapters(mangaId)
                 // Sort the chapters according to current sort direction
                 val sortedChapters = sortChapters(chapters, _uiState.value.chapterSortAscending)
                 _uiState.update { it.copy(chapters = sortedChapters, chaptersLoading = false) }
             } catch (e: Exception) {
-                // Error handling remains the same
                 handleChaptersFetchError(e)
             }
         }
@@ -291,312 +251,6 @@ class MangaDetailedViewModel(
             // In descending order, previous chapter is the next index
             chapterNumbers.getOrNull(currentIdx + 1)
         }
-    }
-
-    private suspend fun fetchMangaFromAPI(id: String): Manga? {
-        return withContext(Dispatchers.IO) {
-            try {
-                val url = "https://api.mangadex.org/manga/$id?includes[]=cover_art&includes[]=author&includes[]=artist&includes[]=tag"
-                val connection = URL(url).openConnection() as HttpURLConnection
-                connection.requestMethod = "GET"
-                connection.connectTimeout = 15000
-                connection.readTimeout = 15000
-
-                if (connection.responseCode == 200) {
-                    val response = connection.inputStream.bufferedReader().use { it.readText() }
-                    val jsonObject = JSONObject(response)
-                    val data = jsonObject.getJSONObject("data")
-                    val attributes = data.getJSONObject("attributes")
-
-                    // Extract title
-                    val titleObj = attributes.getJSONObject("title")
-                    val titles = mutableListOf<String>()
-                    for (key in titleObj.keys()) {
-                        titleObj.getString(key)?.let { titles.add(it) }
-                    }
-
-                    // Extract description
-                    val description = attributes.optJSONObject("description")?.let { descObj ->
-                        descObj.optString("en") ?: run {
-                            // If English description not available, get the first available description
-                            for (key in descObj.keys()) {
-                                val desc = descObj.optString(key)
-                                if (desc.isNotEmpty()) return@run desc
-                            }
-                            ""
-                        }
-                    } ?: ""
-
-                    val status = attributes.optString("status", "unknown")
-
-                    val year = attributes.optInt("year", 0)
-
-                    attributes.optString("contentRating", "safe")
-
-                    var coverImage: String? = null
-                    val relationships = data.getJSONArray("relationships")
-                    for (i in 0 until relationships.length()) {
-                        val rel = relationships.getJSONObject(i)
-                        if (rel.getString("type") == "cover_art") {
-                            if (rel.has("attributes")) {
-                                val coverAttributes = rel.getJSONObject("attributes")
-                                val fileName = coverAttributes.optString("fileName")
-                                if (fileName.isNotEmpty()) {
-                                    coverImage = "https://uploads.mangadex.org/covers/$id/$fileName"
-                                }
-                            }
-                        }
-                    }
-
-                    // Extract tags
-                    val tagsArray = attributes.getJSONArray("tags")
-                    val tags = mutableListOf<SearchViewModel.MangaTag>()
-                    for (i in 0 until tagsArray.length()) {
-                        val tag = tagsArray.getJSONObject(i)
-                        val tagId = tag.getString("id")
-
-                        // Extract tag name from attributes
-                        val tagAttributes = tag.getJSONObject("attributes")
-                        val tagNameObj = tagAttributes.getJSONObject("name")
-
-                        // Try to get English name first, then fallback to any available language
-                        val tagName = tagNameObj.optString("en") ?: run {
-                            var name = ""
-                            val keys = tagNameObj.keys()
-                            while (keys.hasNext()) {
-                                val key = keys.next()
-                                val localizedName = tagNameObj.optString(key)
-                                if (localizedName.isNotEmpty()) {
-                                    name = localizedName
-                                    break
-                                }
-                            }
-                            name
-                        }
-
-                        // Only add tags with non-empty names
-                        if (tagName.isNotEmpty()) {
-                            Log.d(TAG, "Adding tag: $tagName with id: $tagId")
-                            tags.add(SearchViewModel.MangaTag(tagId, tagName))
-                        }
-                    }
-
-                    return@withContext Manga(
-                        id = id,
-                        title = titles,
-                        poster = coverImage,
-                        status = status,
-                        description = description,
-                        lastUpdated = attributes.optString("updatedAt"),
-                        lastChapter = null,
-                        year = if (year > 0) year else null,
-                        tags = tags,
-                        latestUploadedChapterId = attributes.optString("latestUploadedChapter"),
-                        contentRating = attributes.optString("contentRating", null)
-                    )
-                }
-
-                // Handle error responses
-                val errorMessage = when(connection.responseCode) {
-                    404 -> "Manga not found"
-                    429 -> "Rate limit exceeded. Please try again later."
-                    500, 502, 503, 504 -> "Server error. Please try again later."
-                    else -> "Error ${connection.responseCode}: ${connection.responseMessage}"
-                }
-
-                Log.e(TAG, "API error: $errorMessage")
-                throw IOException(errorMessage)
-            } catch (e: Exception) {
-                Log.e(TAG, "Error parsing manga details", e)
-                throw e
-            }
-        }
-    }
-
-    private suspend fun fetchChaptersFromAPI(mangaId: String): List<ChapterModel> {
-        return withContext(Dispatchers.IO) {
-            try {
-                // Get all languages but keep them separated for display
-                val url = "https://api.mangadex.org/manga/$mangaId/feed?limit=500&order[chapter]=asc&includes[]=scanlation_group"
-                val connection = URL(url).openConnection() as HttpURLConnection
-                connection.requestMethod = "GET"
-                connection.connectTimeout = 15000
-                connection.readTimeout = 15000
-
-                if (connection.responseCode == 200) {
-                    val response = connection.inputStream.bufferedReader().use { it.readText() }
-                    val jsonObject = JSONObject(response)
-                    val dataArray = jsonObject.getJSONArray("data")
-                    val chapters = mutableListOf<ChapterModel>()
-                    val readChapters = getReadChapters(mangaId)
-
-                    // Map to store chapter number to its translations
-                    val chapterMap = mutableMapOf<String, MutableList<ChapterModel>>()
-
-                    for (i in 0 until dataArray.length()) {
-                        val chapterObject = dataArray.getJSONObject(i)
-                        val chapterId = chapterObject.getString("id")
-                        val attributes = chapterObject.getJSONObject("attributes")
-
-                        val chapterNumber = attributes.optString("chapter", "")
-                        if (chapterNumber.isEmpty()) continue  // Skip chapters without numbers
-
-                        // Extract volume information
-                        val volume = attributes.optString("volume", null)
-
-                        // Get language code
-                        val language = attributes.optString("translatedLanguage", "en")
-
-                        // Get language flag emoji
-                        val languageFlag = getLanguageFlag(language)
-
-                        // Get scanlation group
-                        var translatorGroup: String? = null
-                        val relationships = chapterObject.getJSONArray("relationships")
-                        for (j in 0 until relationships.length()) {
-                            val rel = relationships.getJSONObject(j)
-                            if (rel.getString("type") == "scanlation_group" && rel.has("attributes")) {
-                                translatorGroup = rel.getJSONObject("attributes").optString("name")
-                                break
-                            }
-                        }
-
-                        val chapterTitle = if (attributes.isNull("title")) null else attributes.getString("title")
-                        val finalTitle = if (chapterTitle.isNullOrBlank()) {
-                            "Chapter $chapterNumber"
-                        } else {
-                            chapterTitle
-                        }
-
-                        val chapterModel = ChapterModel(
-                            id = chapterId,
-                            number = chapterNumber,
-                            title = finalTitle,
-                            publishedAt = attributes.optString("publishAt", ""),
-                            pages = attributes.optInt("pages", 0),
-                            volume = volume,
-                            isRead = readChapters.contains(chapterId),
-                            language = language,
-                            translatorGroup = translatorGroup,
-                            languageFlag = languageFlag
-                        )
-
-                        // Add to chapter map
-                        if (!chapterMap.containsKey(chapterNumber)) {
-                            chapterMap[chapterNumber] = mutableListOf()
-                        }
-                        chapterMap[chapterNumber]?.add(chapterModel)
-                    }
-
-                    // Sort each chapter's translations by language (English first)
-                    chapterMap.forEach { (_, translations) ->
-                        translations.sortWith(compareBy<ChapterModel> { it.language != "en" }
-                            .thenBy { it.publishedAt }
-                            .thenBy { it.language })
-                    }
-
-                    // Flatten the map back to a list, sorted by chapter number
-                    val sortedChapterNumbers = chapterMap.keys
-                        .sortedWith(compareBy<String> {
-                            it.toFloatOrNull() ?: Float.MAX_VALUE
-                        })
-
-                    for (chapterNum in sortedChapterNumbers) {
-                        chapters.addAll(chapterMap[chapterNum] ?: emptyList())
-                    }
-
-                    return@withContext chapters
-                }
-
-                val errorMessage = when(connection.responseCode) {
-                    404 -> "Chapters not found"
-                    429 -> "Rate limit exceeded. Please try again later."
-                    500, 502, 503, 504 -> "Server error. Please try again later."
-                    else -> "Error ${connection.responseCode}: ${connection.responseMessage}"
-                }
-
-                Log.e(TAG, "API error: $errorMessage")
-                throw IOException(errorMessage)
-            } catch (e: Exception) {
-                Log.e(TAG, "Error fetching chapters", e)
-                throw e
-            }
-        }
-    }
-
-    private fun getLanguageFlag(languageCode: String): String {
-        // Extract the base language and region if present (e.g., "en-us" -> "en" and "us")
-        val parts = languageCode.split("-")
-        if (parts.size > 1) parts[1].uppercase() else ""
-
-        // For languages with specific regions, use the region's flag
-        return when (languageCode) {
-            "en" -> "🇬🇧"  // English (default to UK)
-            "en-us" -> "🇺🇸"  // English (US)
-            "ja" -> "🇯🇵"  // Japanese
-            "ko" -> "🇰🇷"  // Korean
-            "zh" -> "🇨🇳"  // Chinese (simplified)
-            "zh-hk", "zh-tw" -> "🇹🇼"  // Chinese (traditional)
-            "ru" -> "🇷🇺"  // Russian
-            "fr" -> "🇫🇷"  // French
-            "de" -> "🇩🇪"  // German
-            "es" -> "🇪🇸"  // Spanish
-            "pt-br" -> "🇧🇷"  // Portuguese (Brazil)
-            "pt" -> "🇵🇹"  // Portuguese
-            "it" -> "🇮🇹"  // Italian
-            "pl" -> "🇵🇱"  // Polish
-            "tr" -> "🇹🇷"  // Turkish
-            "th" -> "🇹🇭"  // Thai
-            "vi" -> "🇻🇳"  // Vietnamese
-            "ar" -> "🇸🇦"  // Arabic
-            "id" -> "🇮🇩"  // Indonesian
-            "fi" -> "🇫🇮"  // Finnish
-            "da" -> "🇩🇰"  // Danish
-            "no" -> "🇳🇴"  // Norwegian
-            "sv" -> "🇸🇪"  // Swedish
-            "nl" -> "🇳🇱"  // Dutch
-            "cs" -> "🇨🇿"  // Czech
-            "hu" -> "🇭🇺"  // Hungarian
-            "ro" -> "🇷🇴"  // Romanian
-            "sk" -> "🇸🇰"  // Slovak
-            "bg" -> "🇧🇬"  // Bulgarian
-            "hr" -> "🇭🇷"  // Croatian
-            "lt" -> "🇱🇹"  // Lithuanian
-            "lv" -> "🇱🇻"  // Latvian
-            "et" -> "🇪🇪"  // Estonian
-            "sl" -> "🇸🇮"  // Slovenian
-            "el" -> "🇬🇷"  // Greek
-            "iw" -> "🇮🇱"  // Hebrew
-            "ms" -> "🇲🇾"  // Malay
-            "sw" -> "🇰🇪"  // Swahili
-            "tl" -> "🇵🇭"  // Filipino
-            "uk" -> "🇺🇦"  // Ukrainian
-            "hi" -> "🇮🇳"  // Hindi
-            "bn" -> "🇧🇩"  // Bengali
-            "pa" -> "🇵🇰"  // Punjabi
-            "is" -> "🇮🇸"  // Icelandic
-            "ga" -> "🇮🇪"  // Irish
-            "cy" -> "🇬🇧"  // Welsh
-            "eu" -> "🇪🇸"  // Basque
-            "ca" -> "🇪🇸"  // Catalan
-            "sw" -> "🇰🇪"  // Swahili
-            "am" -> "🇪🇹"  // Amharic
-            "fa" -> "🇮🇷"  // Persian
-            "zu" -> "🇿🇦"  // Zulu
-            "xh" -> "🇿🇦"  // Xhosa
-            "af" -> "🇿🇦"  // Afrikaans
-            "sq" -> "🇦🇱"  // Albanian
-            "mk" -> "🇲🇰"  // Macedonian
-            "sr" -> "🇷🇸"  // Serbian
-            "bs" -> "🇧🇦"  // Bosnian
-            "tl" -> "🇵🇭"  // Tagalog
-            "jw" -> "🇮🇩"  // Javanese
-            else -> "🌐"  // Default - globe for unknown languages
-        }
-    }
-
-    private suspend fun getReadChapters(mangaId: String): Set<String> {
-        return emptySet()
     }
 
     fun toggleFavorite() {
